+++ crates/deputy-clients/src/golang/cache.rs
@@ -1,18 +1,22 @@
 use crate::shared::{RequestCacheMap, RequestResult};
 
-use super::models::ModuleVersion;
+use super::models::{Module, Package, SearchResults, Versions};
 
 #[derive(Debug, Clone)]
 pub(super) struct GolangCache {
-    pub version_lists: RequestCacheMap<RequestResult<Vec<String>>>,
-    pub latest_versions: RequestCacheMap<RequestResult<ModuleVersion>>,
+    pub modules: RequestCacheMap<RequestResult<Module>>,
+    pub packages: RequestCacheMap<RequestResult<Package>>,
+    pub searches: RequestCacheMap<RequestResult<SearchResults>>,
+    pub versions: RequestCacheMap<RequestResult<Versions>>,
 }
 
 impl GolangCache {
     pub fn new() -> Self {
         Self {
-            version_lists: RequestCacheMap::new_mins(60, 15),
-            latest_versions: RequestCacheMap::new_mins(60, 15),
+            modules: RequestCacheMap::new_mins(60, 15),
+            packages: RequestCacheMap::new_mins(60, 15),
+            searches: RequestCacheMap::new_mins(60, 15),
+            versions: RequestCacheMap::new_mins(60, 15),
         }
     }
 }