+++ crates/deputy-clients/src/golang/mod.rs
@@ -2,12 +2,9 @@ use tracing::error;
 
 use crate::shared::{Request, RequestResult};
 
-use super::github::GithubClient;
-
 mod cache;
 mod consts;
 mod requests;
-mod util;
 
 pub mod models;
 
@@ -16,15 +13,13 @@ use self::cache::GolangCache;
 #[derive(Debug, Clone)]
 pub struct GolangClient {
     cache: GolangCache,
-    github: GithubClient,
 }
 
 impl GolangClient {
     #[must_use]
-    pub fn new(github: GithubClient) -> Self {
+    pub fn new() -> Self {
         Self {
             cache: GolangCache::new(),
-            github,
         }
     }
 
@@ -41,6 +36,6 @@ impl GolangClient {
 
 impl Default for GolangClient {
     fn default() -> Self {
-        Self::new(GithubClient::new())
+        Self::new()
     }
 }